refactor: fix more warnings with lints

+++ merk/src/tree/encoding.rs
@@ -204,8 +204,8 @@ mod tests {
         }) = tree.link(true)
         {
             assert_eq!(*key, [2]);
-            assert_eq!(*child_heights, (123 as u8, 124 as u8));
-            assert_eq!(*hash, [66 as u8; 32]);
+            assert_eq!(*child_heights, (123u8, 124u8));
+            assert_eq!(*hash, [66u8; 32]);
         } else {
             panic!("Expected Link::Reference");
         }
